Fix picks and set operations don't play together well
If a pick is done instead of a full projection, data is kept and only
the schema is altered which is good for performance reasons. However,
set operations such as set difference or union require the data to be
pruned. We now do that lazily only if needed through the
coalesce_helper.