Implement negation through anti joins instead of set differences.
This lifts the previous schema equality restriction between the negated
and positive part of a rule.
Furthermore, the mvr store without a causal broadcast is now another test case.
The test now checks for the issues arising from omitting the causal broadcast
in both cases of the MVR CRDT key value store.